Fix STL mesh reading.
I wasn't checking for the end of the file in my loop properly.
I was using

> while (!input_file.eof())

which is actually wrong, see:
https://stackoverflow.com/questions/5605125/why-is-iostreameof-inside-a-loop-condition-i-e-while-stream-eof-cons

instead I read number of triangles from the header, assert that the
file the expected size for that number of triangles, and then read
exactly that many triangles and close the file. I also turn on the
exception for that stream, which would have caught this originally.